joplin-ml-assistant/requirements.txt
markdown
sentence-transformers
faiss-cpu
llama-cpp-python
streamlit
requests
